pages/dashboard.py
import streamlit as st
import cv2
import numpy as np
from datetime import datetime

st.set_page_config(page_title="Dashboard - SmartFlow", page_icon="📊")

st.title("Traffic Monitoring Dashboard")
st.markdown("Real-time traffic monitoring and signal optimization")

# Placeholder for video feed
st.subheader("Live Traffic Feed")
video_placeholder = st.empty()
st.info("Video feed will be implemented here")

# Traffic statistics
col1, col2, col3 = st.columns(3)
with col1:
    st.metric(label="Current Traffic Density", value="Medium", delta="+5%")
with col2:
    st.metric(label="Average Wait Time", value="45s", delta="-10%")
with col3:
    st.metric(label="Signal Efficiency", value="85%", delta="+2%")

# Traffic flow chart placeholder
st.subheader("Traffic Flow Analysis")
chart_placeholder = st.empty()
st.info("Traffic flow charts will be implemented here")

# System status
st.subheader("System Status")
status_col1, status_col2 = st.columns(2)
with status_col1:
    st.info("Camera Status: Active")
    st.info("AI Model Status: Running")
with status_col2:
    st.info("Last Update: " + datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
    st.info("System Health: Good") 